# .gnva-sauvegardes/rapports-2026-10-04T05-53-07-902Z/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: API GNVA, autorisations territoriales et opérations transactionnelles
servers:
  - url: /api
paths:
...
  /v1/auth/mot-de-passe:
    post:
      operationId: changerMotDePasse
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangementMotDePasse" }
      responses:
        "200":
          description: Modifié
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Resultat" }
...
  /v1/attributions:
    post:
      operationId: attribuerAutocollant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AttributionInput" }
      responses:
        "201":
          description: Attribution atomique
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Donnees" }
...
  /v1/recouvrements/{id}/validation:
    post:
      operationId: validerRecouvrement
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransitionInput" }
      responses:
        "200":
          description: Validation enregistrée
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Donnees" }
...
components:
  schemas:
...
    Donnees:
      type: object
      additionalProperties: true
    Resultat:
      type: object
      required: [message]
      properties:
        message: { type: string }
        nombre: { type: integer }
...
    ChangementMotDePasse:
      type: object
      required: [actuel, nouveau]
      properties:
        actuel: { type: string }
        nouveau: { type: string, minLength: 12 }
...
    AttributionInput:
      type: object
      required: [assujettiId, numeroAutocollant, numeroTimbre]
      properties:
        assujettiId: { type: string }
        numeroAutocollant: { type: string }
        numeroTimbre: { type: string }
...
    TransitionInput:
      type: object
      required: [statut]
      properties:
        statut: { type: string }
        commentaire: { type: string }